requirements.txt
# ==============================================================================
# ADG Knowledge Management System - Python Dependencies
# ==============================================================================

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# Configuration & Validation
pydantic>=2.6.0
pydantic-settings>=2.0.0

# Google APIs
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0

# NotebookLM Integration
notebooklm-mcp

# Background Tasks & Scheduling
apscheduler>=3.10.0

# Server-Sent Events
sse-starlette>=1.6.0

# HTML Templates (for future use)
jinja2>=3.1.0
aiofiles>=23.0.0

# Security (for future authentication)
# python-jose[cryptography]>=3.3.0
# passlib[bcrypt]>=1.7.4

# Database (for future document management)
# sqlalchemy>=2.0.0
# asyncpg>=0.29.0
# alembic>=1.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
